Test-DockerCompose/requirements.txt
click==7.1.2
Flask==2.0.1
Flask-CORS==3.0.10
psycopg2-binary==2.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Werkzeug==2.0.1
bcrypt==3.2.0
